requirements.txt
Flask
python-dotenv
sqlalchemy 
flask_sqlalchemy
psycopg2-binary
